rename and deprint
