+++ docs/api.md
@@ -1,9 +1,9 @@
 ::: qcmanybody.ManyBodyCalculator
 
-::: qcmanybody.qcng_computer.ManyBodyComputerQCNG
+::: qcmanybody.qcng_computer.ManyBodyComputer
     options:
         show_root_heading: true
         inherited_members: true
         members: true
 
-$pydantic: qcmanybody.qcng_computer.ManyBodyComputerQCNG
+$pydantic: qcmanybody.qcng_computer.ManyBodyComputer
