+++ qcmanybody/models/manybody_input_pydv1.py
@@ -171,10 +171,10 @@ class ManyBodyKeywords(ProtoModel):
         # definitive description
         description="Target the supersystem total/interaction energy (IE) data over the many-body expansion (MBE) "
             "analysis, thereby omitting intermediate-body calculations. When False (default), compute each n-body level "
-            "in the MBE up through ``max_nbody``. When True (only allowed for ``max_nbody = nfragments``\), only compute "
+            "in the MBE up through ``max_nbody``. When True (only allowed for ``max_nbody = nfragments`` ), only compute "
             "enough for the overall interaction/total energy: max_nbody-body and 1-body. When True, properties "
             "``INTERACTION {driver} THROUGH {max_nbody}-BODY`` will always be available; "
-            "``TOTAL {driver} THROUGH {max_nbody}-BODY`` will be available depending on ``return_total_data``\; and "
+            "``TOTAL {driver} THROUGH {max_nbody}-BODY`` will be available depending on ``return_total_data`` ; and "
             "``{max_nbody}-BODY CONTRIBUTION TO {driver}`` won't be available (except for dimers). This keyword produces "
             "no savings for a two-fragment molecule. But for the interaction energy of a three-fragment molecule, for "
             "example, 2-body subsystems can be skipped with ``supersystem_ie_only=True``. Do not use with ``vmfc`` in "
@@ -201,7 +201,7 @@ class ManyBodySpecification(ProtoModel):
     schema_name: Literal["qcschema_manybodyspecification"] = "qcschema_manybodyspecification"
     #provenance: Provenance = Field(Provenance(**provenance_stamp(__name__)), description=Provenance.__doc__)
     keywords: ManyBodyKeywords = Field(..., description=ManyBodyKeywords.__doc__)
-    #program: str = Field(..., description="The program for which the Specification is intended.")
+    #program: str = Field(..., description="The program for which the Specification is intended.")  # TODO is qcmanybody
     protocols: ManyBodyProtocols = Field(ManyBodyProtocols(), description=str(ManyBodyProtocols.__doc__))
     driver: DriverEnum = Field(
         ...,
